requirements.txt
asgiref==3.7.2
Django==5.0.2
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
djangorestframework-simplejwt==5.3.1
inflection==0.5.1
mysql==0.0.3
mysqlclient==2.2.4
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2024.1
